requirements.txt
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0